+++ src/main/java/lotto/domain/Lotto.java
@@ -1,6 +1,7 @@
-package lotto;
+package lotto.domain;
 
 import java.util.List;
+import lotto.Rank;
 
 public class Lotto {
     private final List<Integer> numbers;
