+++ src/main/java/lotto/domain/RandomLottoGenerator.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.Collections;
