+++ src/main/java/lotto/domain/WinningNumbers.java
@@ -1,4 +1,6 @@
-package lotto;
+package lotto.domain;
+
+import lotto.Rank;
 
 public record WinningNumbers(Lotto winningLotto, int bonus) {
     public Rank rank(Lotto lotto) {
