+++ src/main/java/lotto/service/LottoMachine.java
@@ -1,7 +1,10 @@
-package lotto;
+package lotto.service;
 
 import java.util.List;
 import java.util.stream.IntStream;
+import lotto.Money;
+import lotto.domain.RandomLottoGenerator;
+import lotto.domain.Lotto;
 
 public class LottoMachine {
     public List<Lotto> makeLottosWith(Money money) {
