+++ src/test/java/lotto/domain/WinningNumbersTest.java
@@ -1,8 +1,11 @@
-package lotto;
+package lotto.domain;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-import java.util.List;
+import lotto.LottoFixture;
+import lotto.Rank;
+import lotto.domain.Lotto;
+import lotto.domain.WinningNumbers;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
