+++ src/test/java/lotto/service/LottoServiceTest.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.service;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.BDDMockito.given;
@@ -6,6 +6,11 @@ import static org.mockito.Mockito.mock;
 
 import java.util.EnumMap;
 import java.util.List;
+import lotto.LottoFixture;
+import lotto.Money;
+import lotto.Rank;
+import lotto.domain.Lotto;
+import lotto.domain.WinningNumbers;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
